Invite Page (#8)
* feat: add updates page

This page is used when the email is sent which contains updates to an event created/attended by user

* feat: add validation for email body

* fix: updatesText for update.html

* fix: remove name and add event for the email, change type of bcc to list[email]

* Add RSVP invite email functionality and template

Introduced a new RSVP invite functionality by adding support for a new template in `app/config.py` and `app/schemas/email.py`. A dedicated HTML template `rsvp/invite.html` was created for rendering invite emails with required placeholders, and validation logic was updated to ensure required fields are present for invites.

* Revert "Add RSVP invite email functionality and template"

This reverts commit 6324f108a415d7cd8bf1838c0778a888493d705a.

* Reapply "Add RSVP invite email functionality and template"

This reverts commit fc0e7411a97f89fd10b8442ed0d0a72f1049e2b2.

* fix: remove .idea files